ML_Models/pothole_detection/pothole_detection/requirements.txt
fastapi
uvicorn
pillow
ultralytics
numpy
opencv-python
torch
torchvision
pydantic
python-multipart
pandas
tqdm
seaborn